separated config package
