+++ cmd/configuration/config.go
@@ -1,4 +1,4 @@
-package main
+package configuration
 
 import (
 	"github.com/KyberNetwork/reserve-data/blockchain"
