+++ cmd/configuration/kovan.go
@@ -1,4 +1,4 @@
-package main
+package configuration
 
 import (
 	"log"
@@ -7,13 +7,12 @@ import (
 	"github.com/KyberNetwork/reserve-data/common"
 	"github.com/KyberNetwork/reserve-data/data/fetcher"
 	"github.com/KyberNetwork/reserve-data/data/storage"
-	"github.com/KyberNetwork/reserve-data/http"
 	"github.com/KyberNetwork/reserve-data/signer"
 	ethereum "github.com/ethereum/go-ethereum/common"
 )
 
-func GetConfigForMainnet() *Config {
-	settingPath := "/go/src/github.com/KyberNetwork/reserve-data/cmd/mainnet_setting.json"
+func GetConfigForKovan() *Config {
+	settingPath := "/go/src/github.com/KyberNetwork/reserve-data/cmd/kovan_setting.json"
 	addressConfig, err := common.GetAddressConfigFromFile(settingPath)
 	if err != nil {
 		log.Fatalf("Config file %s is not found. Error: %s", settingPath, err)
@@ -24,8 +23,8 @@ func GetConfigForMainnet() *Config {
 		log.Fatalf("Fees file cannot found at: %s", feePath, err)
 	}
 	wrapperAddr := ethereum.HexToAddress(addressConfig.Wrapper)
-	reserveAddr := ethereum.HexToAddress(addressConfig.Reserve)
 	pricingAddr := ethereum.HexToAddress(addressConfig.Pricing)
+	reserveAddr := ethereum.HexToAddress(addressConfig.Reserve)
 	burnerAddr := ethereum.HexToAddress(addressConfig.FeeBurner)
 	networkAddr := ethereum.HexToAddress(addressConfig.Network)
 
@@ -39,34 +38,22 @@ func GetConfigForMainnet() *Config {
 		tokens = append(tokens, tok)
 	}
 
-	storage, err := storage.NewBoltStorage("/go/src/github.com/KyberNetwork/reserve-data/cmd/mainnet.db")
+	storage, err := storage.NewBoltStorage("/go/src/github.com/KyberNetwork/reserve-data/cmd/kovan.db")
 	if err != nil {
 		panic(err)
 	}
 
 	fetcherRunner := fetcher.NewTickerRunner(3*time.Second, 2*time.Second, 3*time.Second, 5*time.Second, 5*time.Second)
 
-	fileSigner, depositSigner := signer.NewFileSigner("/go/src/github.com/KyberNetwork/reserve-data/cmd/mainnet_config.json")
+	fileSigner, depositSigner := signer.NewFileSigner("/go/src/github.com/KyberNetwork/reserve-data/cmd/config.json")
 
-	exchangePool := NewMainnetExchangePool(
+	exchangePool := NewKovanExchangePool(
 		feeConfig, addressConfig, fileSigner, storage,
 	)
 
-	hmac512auth := http.KNAuthentication{
-		fileSigner.KNSecret,
-		fileSigner.KNReadOnly,
-		fileSigner.KNConfiguration,
-		fileSigner.KNConfirmConf,
-	}
-
-	endpoint := "https://mainnet.infura.io"
-	bkendpoints := []string{
-		"https://node.kyber.network",
-		"https://mainnet.infura.io",
-		"https://api.mycryptoapi.com/eth",
-		"https://api.myetherapi.com/eth",
-		"https://mew.giveth.io/",
-	}
+	// endpoint := "http://localhost:8545"
+	endpoint := "https://kovan.infura.io"
+	bkendpoints := []string{}
 
 	return &Config{
 		ActivityStorage:         storage,
@@ -78,8 +65,6 @@ func GetConfigForMainnet() *Config {
 		Exchanges:               exchangePool.CoreExchanges(),
 		BlockchainSigner:        fileSigner,
 		DepositSigner:           depositSigner,
-		EnableAuthentication:    true,
-		AuthEngine:              hmac512auth,
 		EthereumEndpoint:        endpoint,
 		BackupEthereumEndpoints: bkendpoints,
 		SupportedTokens:         tokens,
